Closes #301: Profile icons now use getIcon() API. Overrides now possible:
For themes to override, create a plugin hook listening to 'entity:icon:url' and object 'user'.

In the hook return a different url.

To replace default user icons in a plugin one might create a hook:

function profile_usericon_hook($hook, $entity_type, $returnvalue, $params)
{
	if ((!$returnvalue) && ($params['entity'] instanceof ElggUser))
	{
		// return your default graphic here.
	}
}

And set it to priority 900 (lower priority than the code that displays a pretty icon for users but higher than the default object display code in entities.php)

git-svn-id: https://code.elgg.org/elgg/trunk@2221 36083f99-b078-4883-b0ff-0f9b5a30f544