Change to views.php to fix #5566
This fixes Issue #5566, where your uploaded image for large avatar - before you optionally manually crop it - ends up not being a 200x200 square, even if you gave it a much larger image. According to my testing, if I submitted a big one, I got 200x150. But with this tiny change the initial crop works fine and you get 200x200. I don't know if the master setting on the line below should also have the square parameter set to true, but 550x550 looks like a square to me, so go ahead and do that also here if you know what the master is used for :-)