Update install.rst
Updated to include instructions to install "filter_module" in the httpd.conf of xampp and to explain that failure to install the "filter_module" will cause 500 (Internal Server Error) for every action after the .htaccess file is copied to the site root. This is a problem with xampp and is likely to occur with other apache "distributions" as well.