fix(entities): reverts to 1.8 behavior of ElggEntity->subtype reads
The subtype property was never properly implemented to return the string subtype
but rather returned the integer ID of the subtype. Although this was "fixed" before
this commit, it was decided that it was dangerous to start returning strings where
older plugins might be expecting integers. This deprecates reads of the property
altogether (encouraging the use of getSubtype()) and otherwise returns the behavior
of the property in Elgg 1.8.

Fixes #5920