fix(gatekeeper): stop treating being logged out as an error state
It's quite frequent that someone may get a link to a page or entity
that requires being logged in to view it. Instead of treating this
as an error, this change gently forwards the user to the login
page, giving them a chance to authenticate, after which they are
taken to their original destination.