chore(release): continue even if release branch is already created
With git branch/git checkout the script would fail if you had run
it once already and the release branch had already been created.
This makes it so it ignores the fact that the release branch has
been created and just continues on its merry way, which is a more
intuitive behavior than forcing you to clean up the environment
if the script fails the first time.