chore(tests): Turn off Travis commit message tests
These have proven to be very extremely flaky since they are tied to
particular SHAs which may or may not be available in the git history
depending on how deep of a checkout was made.

We want unit tests to pass or fail depending on the current
state of the project, not the entire history of it.