fix(plugins): trigger plugin hooks when saving plugin settings
As stated in #6820 when the change to setPrivateSetting was introduced it doesn't have in account that calling directly that method doesn't triger the plugin hooks.

In https://github.com/Elgg/Elgg/issues/6820#issuecomment-42271849 @jeabakker explains why ```->``` should be called instead of ```->set``` (deprecated) or ```->setSetting()``` (don't call the plugin hooks).

Fixes #6820