chore(cache): simplifies metadata cache
This optimizes the cache for the most common case of read-only access. We no
longer keep track of whether individual MD names are synchronized with the DB,
nor do we try to inject MD writes into the cache. This greatly simplifies the
hard task of cache invalidation, which was a good hiding place for bugs.

This also segregates the cache by access visibility, which makes the cache
safer to use, and frees us from having to invalidate when ignore access is
toggled.

Fixes #7953