feature(events): allows wider variety of callables to be unregistered
This adds a class usable for matching anonymous functions, dynamic method
callbacks, and invokable objects without the need for an object reference.
The matcher can be passed to the event/hook unregister functions to
successfully remove such types of handlers.

This also renames the event registration test so it’s actively tested by
PHPUnit.

Fixes #7750
Fixes #8069