feature(views): Users can jump directly to content via prev/next links
This adds an option to add a URL fragment to pagination links, jumping
the user directly to the paginated content on pages 2 and after. When
returning to the first page, the fragment is not included, assuming
that the user may actually want to see the full first page.

Should the base_url option contain a fragment, the url_fragment value will
not replace it.