fix(deprecation): visible deprecation errors aren't displayed to admin anymore
We cannot assume all administrators to have the technical skill to upgrade
plugin code or even the capability to install a new plugin version that has
addressed the deprecation issues. Therefore it is best not to display visible
errors at all as they only give the impression that the site is broken even
though this is not the case.