chore(email): Switch to `Zend\Mail` instead of built-in `mail()` function
This lays a good foundation for improving Elgg's flexibility and security
in sending emails. Some of the features we open up because of this change:

 * HTML emails
 * Attachments
 * Improved security (automatic header escaping, etc.)
 * Third-party SMTP services (e.g. Mandrill)
 * Receiving emails
 * In-memory transport for easier testing/development

We aren't introducing our own interfaces because `Zend\Mail`
is plenty well supported and the API is agreeable to us.
Writing a wrapper would just be unnecessary overhead.

BREAKING CHANGE:
We are switching to `Zend\Mail` for sending emails in Elgg 2.0.
It's likely that there are some edge cases that the library
handles differently than Elgg 1.x used to. Take care to test
your email notifications carefully when upgrading to 2.0.

Fixes #5918