chore(assets): Check in assets that don't have bower repos
The whole point of going through composer-asset-plugin is to get automatic updates,
but that's not possible with these 4 libraries because they aren't available on
bower at all, so we get no benefit and only complications from depending on them
via bower-asset/* repos.

This is also much easier than finding a replacement library and updating the code
to use the new library.

Fixes #8420