fix(crypto): Prevents ElggCrypt from creating lower-entropy random bytes
If a server lacks the openssl and mcrypt PHP extensions and cannot read from
/dev/urandom, ElggCrypt uses a slow process measuring microtime() result
values to add entropy.

In this rare case, microtime(true) has been shown to return identical values
for $c1 and $c2. This caused a division by zero warning in the calculation
of $rounds, and $rounds would be set to 0. This likely mildly reduced the
entropy gathered through later iterations of microtime() calls.

This change uses a more precise method to measure the time period between $c1
and $c2 and makes sure it’s never 0.

Fixes #8603