feature(hooks): Adds indication in container permissions hook of checking owner
For some reason ElggEntity::create checks the ability to write to the entity
owner, even if the owner will not be the entity’s container. In this case,
the container_permissions_check hook is called twice, and handlers cannot
tell which $params[‘container’] will actually be the entity container, and
which is really the owner being checked.

This introduces a flag $params[‘checking_owner’] to let handlers know that
$params[‘container’] is actually not the container of the entity being
created.

Fixes #8774