feature(views): Allow sites to specify views.php at root
When you install elgg via composer, you can already use the root directory like a plugin,
but without a manifest.xml and it cannot be disabled. Plugins can use a views.php
in their root to define custom view mappings, so it makes sense for you to use this in
your site's root as well.