feature(hooks): Adds real_container in container permissions hook params
ElggEntity::create checks the ability to write to the entity owner, even
if the owner will not be the entity’s container. Because this results in
two hook triggerings, handlers may be confused about the real container
that will be used.

This adds `$params['real_container']` within the container_permissions_check
hook, which always indicates the real container to be used, even when the
owner is being checked.

Fixes #8774