chore(discussions): Make context of discussions hook-configurable
Adds a hook discussions:allow_context that allows registering a type:subtype
to be available for hosting discussions.

Also allows discussion/owner and view pages to work for hidden groups and non-groups. If the topic container can’t be loaded (group is hidden from the user)
we don’t show the reply form.

Fixes #8757