feature(developers): Always show human-readable translations
Even when the `show_strings` setting is turned on, it's nice to
be able to navigate the app. This makes it so that the translation
keys are just appended to the translated string, rather than replacing
them entirely.

Before: `developers:label:show_strings`

After: `Show raw translation strings (developers:label:show_strings)`

Fixes #8834