feature(relationships): Adds more clear options to relationship queries
Users of elgg_get_entities_from_relationship (and elgg_get_river) can now
explicitly specify whether a relationship GUID given in the query is the
target or the subject of a relationship. When this is done, the function
automatically sets the correct value for "inverse_relationship".

Only in rare cases where a GUID is not given do users still need to consider
the "inverse_relationship" option.

This also rewrites and expands the unit tests for this function and improves
the phpdocs.