chore(messages): unread count again updates when you view new messages
To display an accurate count, the count must be calculated after the new
message is displayed. Previously we were tied to the pagesetup event, but
that now fires early due to the use of resource views. Instead we add the
topbar menu item in the register hook.

BREAKING CHANGE:
If a plugin has removed or replaced messages_notifier to hide/alter the
inbox icon, the plugin must instead do the same for the topbar menu handler
(messages_register_topbar).

Fixes #8862