chore(theme): links inside .elgg-subtext are distinguishable again
The <time> element of comment and discussion_reply objects was earlier
converted to a link and the link color was set to grey. This accidentally
affected all links inside .elgg-subtext. This fix makes the CSS selector
of the <time> links more specific so that regular links still have their
original color.

Fixes #8839