perf(river): no longer needlessly render river responses
For river items that cannot receive comments, we explicitly pass a `responses`
option to the river layout, which bypasses the bulk of river/elements/responses
rendering. Notably a `countComments()` query is no longer performed and an empty
form is no longer rendered for each of these items.

For BC, the `div.elgg-river-responses` wrapper is still created, but contains only
whitespace.

In the river/elements/responses view, we also now correctly check for comments
(and their subclasses like ElggDiscussionReply) so we don't render these. The
core river views do not reach to this point due to changes in this PR, so this
is mainly for plugins that may have overridden those views.

Fixes #9046