fix(likes): likes preloader and entity menus now consider likability
In 2.0 we added the `likes:is_likable` hook to select which entities can receive
likes but we forgot to update the preloader and the entity menu setup to check it.

Now likable entities are preloaded as they should be, and in menus for unlikable
entities we no longer query to generate the likes/count view.

Fixes #9065