fix(a11y): display aalborg mobile site menu toggle in more browsers
The 3 `span.icon-bar` elements do not display in Firefox on Android. This
replaces them with an 88x48 semi-transparent PNG image displayed at 22x12
for clean scaling.

Fixes #9110

BREAKING CHANGE:
Plugins that customized `.elgg-button-nav` (or the spans inside) will need
to be altered. The `.elgg-icon` CSS is left in place but will be removed in
Elgg 2.0.