fix(views): locations specified in /engine/views.php are modifiable
When simplecache is on, all view locations are loaded from cache, so no
dir scanning/configuration should be done. In elgg_views_boot, we'd
forgotten to scan engine/views.php based on this condition. This moves
the scanning to the top and into the system cache conditional block.

This also gets rid of duplicated code that causes a double scan of the
core view files.

Fixes #9308