fix(events): the pagesetup event timing is more like 1.x
In 1.x, the event was triggered just before the first view was rendered
and it was assumed that some global state, particularly context, had been
set by this time in page-specific code.

In 2.0 this assumption is wrong because the page-specific code that sets
state like context often runs *within* a resource view. Hence the event
is triggered too early, before the page-specific state has been set up.

Here we delay the `pagesetup` event if the first view is a resource view.

This allows plugins to move all page-specific logic like context setting
into a resource view with more confidence that that state will be available
in their `pagesetup` event handlers. This likely fixes a class of
undiscovered bugs like #9271 and #9322.

Refs #8374