feature(js): adds elgg/popup AMD module
Converts JS popup to an AMD module subsequently allowing plugins to
manipulate popups programmatically.
Popups can now be bound with .elgg-popup class and data-href attribute in
addition to rel=popup and href.
Popup positioning details can be passed with data-position of the trigger.
Triggers additional 'open' and 'close' jQuery events that can be used to
manipulate the popup element and its contents