fix(errors): nested forward 404 calls are less likely to abruptly fail
Sometimes while drawing the error page, plugin code may request the error
page again (often code within the `pagesetup` event). We no longer fail
with an empty page on the 2nd call since this is more common that expected.

On the 3rd call we assume an infinite loop and just use a Location header
to send the user to the home page.

Fixes #9476