fix(js): again calls widgets.init in the init, system event
In 2.0 the `widgets.init` was called on `init, system`. In 2.1 it was called
whenever the `elgg/widgets` module was loaded, but this introduces the problem
that plugins that worked on 2.0 no longer have a reliable blocking point to
make sure `widgets.init` has been called.

Due to BC, we cannot require plugins depend on the new `elgg/widgets` module,
but we can at least provide assurance that `widgets.init` will have been
called after `init, system`.

Hence, we now call it in `init, system` as before, and we delay `elgg/init`
until `elgg/widgets` has loaded.

Fixes #9523