feature(ajax): better elgg/Ajax handling of form data and URLs
As `jQuery.serialize` is often used to set post data, this change allows a
string form of `options.data` to be passed through. In this case, however,
we do not trigger the request hook, as handlers expect an object.

Because we want requests to be hookable, we nudge devs toward a new `objectify`
method, which "serializes" to an object so it can be passed through the hook.

All methods now accept a query string that gets passed to the underlying URL,
but does not appear in the hook types. `action()` and `path()` accept
absolute site URLs, as one might get from `$anchor.prop('href')`.

Fixes #9534
Fixes #9564