fix(js): deprecate elgg.ui.widgets more reliably
Widgets JS API has been moved to a named elgg/widgets AMD module,
which is inlined in elgg.js to make sure the API is available at runtime.
This allows us to deprecate elgg.ui.widgets in a manner that is less likely
to break plugin code, i.e. ensuring that the API is available when elgg AMD
module is loaded

Fixes #9523