chore(releases): commits can build release notes for the next release
Commits can add notes to the top of `CHANGELOG.md`. When the next release
is built, these notes are placed between the new version header and the
contributors list.

In all tagged releases, changelog will appear as usual, but work
branches may have notes for the next release at the top.

Fixes #9489