feature(js): adds elgg/popup AMD module
Converts JS popup to an AMD module subsequently giving more freedom
to plugins to manipulate popups programmatically.
Popup target module can now be defined with data-href attribute in
addition to href attribute (suitable for use with non-anchor elements).
Popup positioning details can be passed with data-position of the trigger
(which is simpler than using the triggered plugin hook).
jQuery "open" and "close" events are now triggered when the popup is shown
and closed, allowing plugins to manipulate popup module and its contents