fix(discussion): correctly check permissions before showing reply form
Reply form is no longer shown if the viewer can not reply to a discussion
(canWriteToContainer()).

Removes group write permissions check form the discussion view:
a) it was duplicating the logic present in container_permissions_check
hook,
b) an assumption that the user must be allowed to write a discussion to a
group to post a reply is inaccurate (e.g. only group owner might be able
to
create new discussions, while group members might be allowed to post
replies).