feature(core): sites can preview features before the next major release
A user-facing feature can be added to core in a minor release such that it
has to be enabled to be used, but will automatically be enabled in the next
major release.

These options are displayed in Admin > Configuration > Preview Features.

Currently this does not clean up settings for stale features.

Fixes #9711