fix(autoload): fixes bugs in class map saving
The autoloader class map now marks itself altered only if a path is changed or
set initially. This allows `elgg_register_class()` to be called on every request
without triggering a re-save.

In 1.12, the `engine/lib` loader now properly calls the setup function returned
by `autoload.php`, allowing the class map to be saved. This bug does not exist
in 2.x.

Refs #9744