security(actions): don't refresh invalid CSRF tokens in the page
The refresh token process no longer blindly replaces all in-page CSRF tokens,
as these could have been in user content. Only tokens that have been validated
on the server are replaced.

We can validate tokens even if the session had expired because we base the
tokens not on session ID, but rather a unique key stored in the session and
passed to the client.

On token refresh, Elgg now provides a page refresh link and an appropriate
message depending on whether the user was logged out or into another account
(e.g. via another tab).

BREAKING CHANGE:
The JS function `elgg.security.setToken` is now formally marked private and
its parameters are not backwards compatible.