security(passwords): removes all legacy password hashes
In case of a site suffering a DB leak, this removes all remaining legacy
password hashes. Although these were salted, the speed of MD5 makes brute
force attacks on the table data practical.

Users that have logged in under an Elgg 1.10 or later system are not affected.
The others, when logging in, are prompted to reset their password (via email
authentication).

BREAKING CHANGE:
The `password` and `hash` columns are removed from the `users_entity` table
and the attributes are removed from `ElggUser` attributes. The function
`generate_new_password` is also removed.