feature(forms): input/checkbox is now usable with elgg_view_input()
Adds an exception to elgg_view_input() to render the label next to the checkbox input
rather than above the input (as in other input types).
Adds default class for single checkbox inputs wrapped in an html tag with a label
(.elgg-input-single-checkbox)
Adds an option to customize the HTML tag used to wrap the single checkbox input with a label
to avoid double wrapping with a \<label\> tag.

Fixes #9808