chore(tests): eliminates setTimeout race in ElggSpinnerTest
Even though the class should be added at 20ms, with a wait of 100ms tests were
still failing. Apparently there's no way to reliably test that a timeout of
a particular period is used, so now the module fires a test event, allowing
the test to directly listen instead of guessing how long to wait. This still
at least tests that there is *some* delay.