chore(project): separate i18n updates from releases
Language updates are super slow and ruin the DX of cutting releases. Also,
pairing them with releases puts a large amount of pressure on core devs to
rubber stamp language changes without any public review.

This allows language files (code, not docs) to be updated independently of
releases. The follow up to this would be to update the docs in 2.x to recommend
running the language intake separately and sending in standard pull requests
independently of releases.