feat(composer): brings back composer.lock
Success of a build depends on the state of vendor libraries. We need make
sure that vendor libraries in the successful build correspond to the
vendor libraries distributed to end users.

This also adds the benefit of faster Travis builds, Scrutinizer support
(which currently fails due to the composer asset plugin dependency).

Fixes #9430