fix(access): updates no longer mistakenly blocked in some scenarios
In order to fix `canEdit()` in 1.9, `update()` needed to load a fresh copy of
entity from the DB to check the persisted attributes. In that fix,
https://github.com/Elgg/Elgg/commit/66eb9e6d14252314769867cd45aed7557a29532a,
a corner case was checked out of my paranoia that a dev would load an
invisible entity with access control on, but try to delete it with access
off.

I now believe this check was unnecessary (we never did similar checks for
`delete()`). When the original `canEdit()` attributes issue was resolved in