feat(ckeditor): better control over ckeditor initialization and behavior
Individual longtext inputs can now be configured independently.
Adds an option to disalbe an editor for a longtext input.
Adds an option to bind an editor but not activate it until
toggled by the user
Adds jQuery events that can be triggered on the original textarea to
focus or reset the CKEditor instance.
Adds elgg/ckeditor#reset and elgg/ckeditor#focus to reset/focus
programmatically

Fixes #9391