fix(web_services): handle string params with proper escaping
For string params, `serialize_parameters` outputs a string literal of PHP
code, and `addcslashes` was not appropriate for this task, particularly
around escapes in strings. While you can tell it to also escape escape
chars, a second pass would be needed.

The tests now `eval` the output just as the production code does, and
this also makes the tests easier to read.