fix(ajax): elgg/Ajax view() and form() set $vars as expected
The `elgg/Ajax` module was auto-converting any request with `options.data`
to use the `POST` method. This makes sense for things like actions, which
may send a large amount of form data, but it caused `$vars` to not be
populated as expected because only `GET` params are injected.

If you specified `options.data` with these methods before, note the server
will no longer receive this data as `$_POST` on the server.

Fixes #10667