fix(admin): prevents simultaneous plugin (de)activation/reordering
Problems can occur if a plugin order/activation change is requested while
another change is being executed. This places a semi-opaque cover over the
plugins list while an operation is executing so that the user cannot start
another operation.

This will not prevent the problem for users navigating via keyboard.

Fixes #10706