fix(ckeditor): fixes basepath issues on some systems
As the CKeditor library is not a real AMD library, its execution may not be
controlled as well by RequireJS, causing a race condition with the module
that sets the global `CKEDITOR_BASEPATH` constant.

This renames the module `elgg/ckeditor` to `elgg-ckeditor` which eliminates
the need for setting `CKEDITOR_BASEPATH` at all. With the new filename,
the CKeditor library is no longer confused when sniffing for its resources
folder.

Fixes #10724

BREAKING CHANGES:
The AMD module `elgg/ckeditor` is now `elgg-ckeditor`. The module `elgg/ckeditor/set-basepath`
was removed.