chore(boot): allow application to reboot without loosing setups
Adds Includer::requireFileOnce() which requires the file and
stores the returning setup value if it is a Closure.

This allows the tests to replace an instance of an Application
without loosing setup information. This is especially
important with files that should only be required once
(e.g. core files or plugin start.php) because they contain
function declarations as well as hook/event registrations.