feature(widgets): using input/number to select number of displayed items in widget edit views
BREAKING CHANGE:
The view `object/widget/edit/num_display` now uses an `input/number` field instead of an `input/select` field
to set the number of displayed items. Widget edit views might need to be updated if a custom max number
(higher than default_limit or 20) is used or if a custom stepsize of selectable item numbers is wanted.