feat(plugins): separate plugin repository from caching concerns
To simplify testing, plugins DB/repository service no longer manages plugin caching.
Because during testing, upgrades and installation, DB might not be available or plugins
and caches might be in a volatile state, this makes it easier to inject plugins
into related services without having to instantiate the DB and the
plugin repository.

Fixes #12004