doc(install): mention necessary SELinux changes
I noticed that a topic that poster Nikolai Shcherbin seems to address frequently
are questions regarding installing Elgg on CentOS. A lot of people seem to be
getting stuck at various points in the installation procedure, and the common
link appears to be SELinux-related permissions issues. It therefore seems
prescient to document what needs to be done in the official documentation to cut
down on the help requests.